added button to group. Added style to doctor app side. Some adittional fixes to get everything up and running.

+++ app-android-xamarin-Doctor-side/Moodis/Moodis/Feature/SignIn/SignInActivity.cs
@@ -52,7 +52,8 @@ namespace Moodis.Feature.SignIn
             }
             else if (resultCode == Result.Canceled && requestCode == REQUEST_CODE_REGISTER)
             {
-                await DeleteUser();
+                if(SignInViewModel.currentUser != null)
+                    await DeleteUser();
             }
         }
